+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @soramitsu-ui/ui
 
+## 0.13.13
+
+### Patch Changes
+
+- b5029a5b: **feat**(`SSelectDropdown`): add mandatory prop
+
+## 0.13.12
+
+### Patch Changes
+
+- 0a615d5c: **feat**(`STextField`): add validations list prop
+
+## 0.13.11
+
+### Patch Changes
+
+- 8d0d2d66: **fix**(`SDropdown`): add expression to display dropdown options correctly
+
 ## 0.13.10
 
 ### Patch Changes